M  app/Http/Controllers/Address/AddressDestroyController.php M  app/Http/Controllers/Address/AddressStoreController.php M  app/Http/Controllers/Address/AddressUpdateController.php M  routes/api.php

+++ app/Http/Controllers/Address/AddressDestroyController.php
@@ -21,6 +21,11 @@ class AddressDestroyController extends Controller
     {
         $address = Address::findOrFail($id);
 
+        // Проверка на соответствие user_id
+        if ($address->user_id !== $request->user()->id) {
+            return $this->response([], 'У вас нет прав для удаления этого адреса.', JsonResponse::HTTP_FORBIDDEN);
+        }
+
         $address->delete();
 
         return $this->response([], 'Адресс успешно удален!');
